Added typedef decimal_digits_t (uint16) for number of decimals
For fields and Item's uint8 should be good enough. After
discussions with Alexander Barkov we choose uint16 (for now)
as some format functions may accept +256 digits.

The reason for this patch was to make the storage of decimal
digits simlar. Before this patch decimals was stored/used as
uint8, int and uint.

Changed most decimal variables and functions to use the new
typedef.