Avoid mallocs when using LONGLONG_BUFFER_SIZE
When using String::alloc() to allocate a string, the String ensures that
there is space for an extra NULL byte in the buffer and if not, reallocates
the string. This is a problem with the String::set_int() that calls
alloc(LONGLONG_BUFFER_SIZE), which forces an extra malloc/free to
happen.

Fixed by using LONGLONG_BUFFER_SIZE+1 for StringBuffer's.